backend/requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-rest-auth==5.0.2
Django==4.2.11
django-allauth==65.1.0
django-cors-headers==4.9.0
django-filter==25.1
django-phonenumber-field==8.4.0
django-redis==6.0.0
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.3
idna==3.11
packaging==26.0
phonenumbers==9.0.23
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.1
redis==7.1.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
